# apps/api/docs/openapi_v1.0.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CollegeERPAPI",
    "version": "1.0.0",
    "description": "\nCollege ERP System API\n\n**Generated:** 2026-02-03T12:39:45.480081\n\n**Locked Domains:**\n- Admission (v1.0.0) - Contracts locked with strict validation\n\n**Status:**\n- \u2705 Admission: Ready for frontend integration\n- \u23f3 Finance: In progress\n- \u23f3 Academic: In progress\n- \u23f3 Student: In progress\n\n**Breaking Changes Policy:**\n- Enum additions: Safe (backward compatible)\n- New optional fields: Safe\n- Required field changes: Requires migration + deprecation notice\n- Enum removals: 6-month deprecation period\n- Field type changes: Major version bump\n        "
  },
  "paths": {
...
    "/api/v1/campus/transport/transport/vehicles": {
      "post": {
        "tags": [
          "campus",
          "Campus - Transport",
          "Transport Logistics"
        ],
        "summary": "Create Vehicle",
        "description": "Add a new vehicle to the transport fleet",
        "operationId": "create_vehicle_api_v1_campus_transport_transport_vehicles_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "campus",
          "Campus - Transport",
          "Transport Logistics"
        ],
        "summary": "List Vehicles",
        "description": "List vehicles in the fleet",
        "operationId": "list_vehicles_api_v1_campus_transport_transport_vehicles_get",
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  },
                  "title": "Response List Vehicles Api V1 Campus Transport Transport Vehicles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/transport/transport/routes": {
      "get": {
        "tags": [
          "campus",
          "Campus - Transport",
          "Transport Logistics"
        ],
        "summary": "List Routes",
        "description": "List all transport routes",
        "operationId": "list_routes_api_v1_campus_transport_transport_routes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransportRoute"
                  },
                  "type": "array",
                  "title": "Response List Routes Api V1 Campus Transport Transport Routes Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "campus",
          "Campus - Transport",
          "Transport Logistics"
        ],
        "summary": "Create Route",
        "description": "Define a new transport route",
        "operationId": "create_route_api_v1_campus_transport_transport_routes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRoute"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/campus/transport/transport/allocations": {
      "post": {
        "tags": [
          "campus",
          "Campus - Transport",
          "Transport Logistics"
        ],
        "summary": "Allocate Transport",
        "description": "Assign a student to a transport route and vehicle",
        "operationId": "allocate_transport_api_v1_campus_transport_transport_allocations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportAllocationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportAllocation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "campus",
          "Campus - Transport",
          "Transport Logistics"
        ],
        "summary": "List Allocations",
        "description": "List transport allocations",
        "operationId": "list_allocations_api_v1_campus_transport_transport_allocations_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "name": "route_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransportAllocation"
                  },
                  "title": "Response List Allocations Api V1 Campus Transport Transport Allocations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/transport/transport/tracking": {
      "post": {
        "tags": [
          "campus",
          "Campus - Transport",
          "Transport Logistics"
        ],
        "summary": "Log Gps Data",
        "description": "Record GPS location for a vehicle",
        "operationId": "log_gps_data_api_v1_campus_transport_transport_tracking_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GPSData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleGPSLog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/transport/transport/statistics/summary": {
      "get": {
        "tags": [
          "campus",
          "Campus - Transport",
          "Transport Logistics"
        ],
        "summary": "Get Fleet Statistics",
        "description": "Get overall transport statistics",
        "operationId": "get_fleet_statistics_api_v1_campus_transport_transport_statistics_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GPSData": {
        "properties": {
          "vehicle_id": {
            "type": "integer",
            "title": "Vehicle Id"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "latitude",
          "longitude"
        ],
        "title": "GPSData"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RouteCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "start_point": {
            "type": "string",
            "title": "Start Point"
          },
          "end_point": {
            "type": "string",
            "title": "End Point"
          },
          "fare_amount": {
            "type": "number",
            "title": "Fare Amount"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code",
          "start_point",
          "end_point",
          "fare_amount"
        ],
        "title": "RouteCreate"
      },
...
      "TransportAllocation": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "route_id": {
            "type": "integer",
            "title": "Route Id"
          },
          "vehicle_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Id"
          },
          "pickup_point": {
            "type": "string",
            "title": "Pickup Point"
          },
          "drop_point": {
            "type": "string",
            "title": "Drop Point"
          },
          "academic_year": {
            "type": "string",
            "title": "Academic Year"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ACTIVE"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "route_id",
          "pickup_point",
          "drop_point",
          "academic_year"
        ],
        "title": "TransportAllocation",
        "description": "Commute Assignment - Linking student to logistics"
      },
      "TransportAllocationRequest": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "route_id": {
            "type": "integer",
            "title": "Route Id"
          },
          "pickup_point": {
            "type": "string",
            "title": "Pickup Point"
          },
          "drop_point": {
            "type": "string",
            "title": "Drop Point"
          },
          "academic_year": {
            "type": "string",
            "title": "Academic Year"
          },
          "vehicle_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Id"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "route_id",
          "pickup_point",
          "drop_point",
          "academic_year"
        ],
        "title": "TransportAllocationRequest"
      },
      "TransportRoute": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "start_point": {
            "type": "string",
            "title": "Start Point"
          },
          "end_point": {
            "type": "string",
            "title": "End Point"
          },
          "fare_amount": {
            "type": "number",
            "title": "Fare Amount"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "code",
          "start_point",
          "end_point",
          "fare_amount"
        ],
        "title": "TransportRoute",
        "description": "Logistics Flow - Path and pricing for commute"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Vehicle": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "registration_number": {
            "type": "string",
            "title": "Registration Number"
          },
          "vehicle_type": {
            "type": "string",
            "title": "Vehicle Type"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "registration_number",
          "vehicle_type",
          "capacity"
        ],
        "title": "Vehicle",
        "description": "Transport Fleet - Vehicle Identity ownership"
      },
      "VehicleCreate": {
        "properties": {
          "registration_number": {
            "type": "string",
            "title": "Registration Number"
          },
          "vehicle_type": {
            "type": "string",
            "title": "Vehicle Type"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          }
        },
        "type": "object",
        "required": [
          "registration_number",
          "vehicle_type",
          "capacity"
        ],
        "title": "VehicleCreate"
      },
      "VehicleGPSLog": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "vehicle_id": {
            "type": "integer",
            "title": "Vehicle Id"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "latitude",
          "longitude"
        ],
        "title": "VehicleGPSLog",
        "description": "GPS Telemetry - Real-time vehicle tracking data"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}